[Hue] Implemented fade time
Implemented fade time config for light things.
Implemented ThingAction fadingLightCommand for Hue lights
Updated README.md

Bug: #1722

+++ addons/binding/org.openhab.binding.hue/src/main/java/org/openhab/binding/hue/internal/StateUpdate.java
@@ -187,7 +187,7 @@ public class StateUpdate extends ConfigUpdate {
      * @param timeMillis time in milliseconds [0..6553600]
      * @return this object for chaining calls
      */
-    public StateUpdate setTransitionTime(int timeMillis) {
+    public StateUpdate setTransitionTime(long timeMillis) {
         if (timeMillis < 0 || timeMillis > 6553600) {
             throw new IllegalArgumentException("Transition time out of range");
         }
